Move steps from Release workflow into the Release script [no ci]
Initially, the Release workflow would run on a new version publish. Because we are using Changesets, the version is created via another action, which does not trigger other Actions by default. By moving those steps into the script that Changesets uses, we can simply release to NPM and upload our docs to S3 in one action.